+++ AmazonChimeSDK/AmazonChimeSDK/device/DefaultDeviceController.swift
@@ -33,7 +33,7 @@ import AVFoundation
 
     public func listAudioDevices() -> [MediaDevice] {
         var inputDevices: [MediaDevice] = []
-        let loudSpeaker = MediaDevice(label: "Built-in Speaker", type: MediaDeviceType.audioBuiltInSpeaker)
+        let loudSpeaker = getSpeakerMediaDevice()
         if let availablePort = audioSession.availableInputs {
             inputDevices = availablePort.map { port in MediaDevice.fromAVSessionPort(port: port) }
         }
@@ -121,5 +121,25 @@ import AVFoundation
 
     public func listSupportedVideoCaptureFormats(mediaDevice: MediaDevice) -> [VideoCaptureFormat] {
         return MediaDevice.listSupportedVideoCaptureFormats(mediaDevice: mediaDevice)
+    }
+
+    public func getActiveAudioDevice() -> MediaDevice? {
+        // Check for speaker case
+        if audioSession.currentRoute.outputs.count > 0 {
+            let currentOutput = audioSession.currentRoute.outputs[0]
+
+            if currentOutput.portType == .builtInSpeaker {
+                return getSpeakerMediaDevice()
+            }
+        }
+        if audioSession.currentRoute.inputs.count > 0 {
+            return MediaDevice.fromAVSessionPort(port: audioSession.currentRoute.inputs[0])
+        } else {
+            return nil
+        }
+    }
+
+    private func getSpeakerMediaDevice() -> MediaDevice {
+        return MediaDevice(label: "Built-in Speaker", type: MediaDeviceType.audioBuiltInSpeaker)
     }
 }